refactor(zero-approval): collect files into module and use ConfirmationPendingContent
