--- apps/cowswap-frontend/src/common/containers/ZeroApprovalModal/ZeroApprovalModal.cosmos.tsx
@@ -1,7 +1,0 @@
-import { ZeroApprovalModal } from './ZeroApprovalModal'
-
-const Fixtures = {
-  default: <ZeroApprovalModal onDismiss={() => alert('dismiss')} />,
-}
-
-export default Fixtures
