+++ apps/cowswap-frontend/src/modules/swap/containers/SwapWidget/index.tsx
@@ -36,9 +36,9 @@ import { TradeWidget, TradeWidgetContainer, useTradePriceImpact } from 'modules/
 import { useTradeRouteContext } from 'modules/trade/hooks/useTradeRouteContext'
 import { useWrappedToken } from 'modules/trade/hooks/useWrappedToken'
 import { useTradeUsdAmounts } from 'modules/usdAmount'
+import { useShouldZeroApprove } from 'modules/zeroApproval'
 
 import { useRateInfoParams } from 'common/hooks/useRateInfoParams'
-import { useShouldZeroApprove } from 'common/hooks/useShouldZeroApprove'
 import { CurrencyInfo } from 'common/pure/CurrencyInputPanel/types'
 import useNativeCurrency from 'lib/hooks/useNativeCurrency'
