+++ apps/cowswap-frontend/src/modules/zeroApproval/hooks/useShouldZeroApprove/shouldZeroApprove.ts
@@ -3,7 +3,7 @@ import { Currency, CurrencyAmount } from '@uniswap/sdk-core'
 
 import { Nullish } from 'types'
 
-import { ApprovalState } from '../useApproveState'
+import { ApprovalState } from 'common/hooks/useApproveState'
 
 interface ShouldZeroApproveBaseParams {
   tokenContract: Nullish<Erc20>
@@ -29,9 +29,10 @@ export async function shouldZeroApprove({
   spender,
   amountToApprove,
   isBundle,
-}: ShouldZeroApproveParams) {
+}: ShouldZeroApproveParams): Promise<boolean> {
   const shouldApprove =
     isBundle || approvalState === ApprovalState.NOT_APPROVED || approvalState === ApprovalState.PENDING
+
   if (!tokenContract || !spender || !amountToApprove || !shouldApprove) {
     return false
   }
