+++ apps/cowswap-frontend/src/modules/zeroApproval/hooks/useShouldZeroApprove/useShouldZeroApprove.ts
@@ -6,11 +6,10 @@ import { Currency, CurrencyAmount } from '@uniswap/sdk-core'
 
 import { Nullish } from 'types'
 
+import { useTradeSpenderAddress } from 'common/hooks/useTradeSpenderAddress'
 import { useApprovalStateForSpender } from 'lib/hooks/useApproval'
 
 import { shouldZeroApprove as shouldZeroApproveFn } from './shouldZeroApprove'
-
-import { useTradeSpenderAddress } from '../useTradeSpenderAddress'
 
 // TODO: Handle tokens that don't allow approvals larger than the balance of the wallet
 export function useShouldZeroApprove(amountToApprove: Nullish<CurrencyAmount<Currency>>): boolean {
