+++ apps/cowswap-frontend/src/modules/zeroApproval/hooks/useZeroApprovalState.ts
@@ -1,0 +1,7 @@
+import { useAtomValue } from 'jotai'
+
+import { zeroApprovalState } from '../state/zeroApprovalState'
+
+export function useZeroApprovalState() {
+  return useAtomValue(zeroApprovalState)
+}
