+++ apps/cowswap-frontend/src/modules/zeroApproval/hooks/useZeroApproveModalState.ts
@@ -1,0 +1,9 @@
+import { ModalState, useModalState } from 'common/hooks/useModalState'
+
+import { useZeroApprovalState } from './useZeroApprovalState'
+
+export function useZeroApproveModalState(): ModalState<void> {
+  const { isApproving } = useZeroApprovalState()
+
+  return useModalState<void>(isApproving)
+}
