+++ apps/cowswap-frontend/src/modules/zeroApproval/index.ts
@@ -1,0 +1,6 @@
+export { shouldZeroApprove, useShouldZeroApprove } from './hooks/useShouldZeroApprove'
+export { useNeedsZeroApproval } from './hooks/useNeedsZeroApproval'
+export { useZeroApprovalState } from './hooks/useZeroApprovalState'
+export { useZeroApprove } from './hooks/useZeroApprove'
+export { useZeroApproveModalState } from './hooks/useZeroApproveModalState'
+export { ZeroApprovalModal } from './containers/ZeroApprovalModal'
